refactor: drop voice_call resource
The "Voice Call" test is deprecated and replaced in favor of "Voice"
aka. "RTP Streams".
